aider: Add "cypress:open" script to package.json.
# Aider chat conversation:

USER: please fix the package.json

ASSISTANT: To fix the `package.json` and add the Cypress open script correctly, I will provide a diff that adds the script within the existing scripts object. Since I don't have the full content of your `package.json`, I will assume that the `scripts` object is not empty and that the `"cypress:open": "cypress open",` line needs to be added correctly.

```diff
--- package.json
+++ package.json
@@ ... @@
   "scripts": {
+    "cypress:open": "cypress open",
     ...
   },
@@ ... @@
```

Please ensure that the `"cypress:open": "cypress open",` line is added within the `scripts` object in your `package.json` file, following the correct JSON syntax. If there are other scripts already present, make sure to place this line in the correct position with a trailing comma if it's not the last item in the object. If you provide the current content of your `package.json` file, I can give you a more accurate diff.

USER: UnifiedDiffNoMatch: hunk failed to apply!

package.json does not contain lines that match the diff you provided!
Try again.
DO NOT skip blank lines, comments, docstrings, etc!
The diff needs to apply cleanly to the lines in package.json!

package.json does not contain these 3 exact lines in a row:
```
  "scripts": {
    ...
  },
```

ASSISTANT: Given the information provided, I will create a diff to add the "cypress:open" script to the `package.json` file. Since I don't have the exact structure of your `package.json`, I will assume that the `scripts` section is not empty and that the new script should be added in a new line within the existing `scripts` object.

```diff
--- package.json
+++ package.json
@@ ... @@
+  "cypress:open": "cypress open",
@@ ... @@
```

Please ensure that the `"cypress:open": "cypress open",` line is added within the `scripts` object in your `package.json` file, following the correct JSON syntax. If there are other scripts already present, make sure to place this line in the correct position with a trailing comma if it's not the last item in the object. If you provide the current content of your `package.json` file, I can give you a more accurate diff.